Add files via upload
This is the audit that I erformed based on the available scenario of how the company was treating the usefull data like PII & SPII at the cost of finance and it's own reputation in the global market. Along this there are some suggestion at the end of the audit which can be followed in order to avoid unnecessary damage to finance and repo.